Refactor: Search component
moved submitSearch logic into Search component
